requirements.txt
asgiref==3.8.1
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
joblib==1.4.2
numpy==2.2.2
opencv-python==4.11.0.86
pillow==11.1.0
scikit-learn==1.6.1
scipy==1.15.1
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1